move GrpoRunConfigPairs into it's own file
